requirements.txt
###### Requirements without Version Specifiers ######
gunicorn

###### Requirements with Version Specifiers ######
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
openai==0.27.0
Markdown==3.4.1
psycopg2-binary==2.9.1
python-dotenv==1.0.0
